pyproject.toml
[project]
name = "discord-slack-relay"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.4",
    "aiohttp>=3.9",
    "python-dotenv>=1.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
